[FE] feat: TOTPModal Input 분리, 포커스 스타일 추가
- TOTPModal 에서 OTPInput 컴포넌트 분리
- OTPInput에 포커스됐을 때 스타일 추가
